requirements/local.txt
-r ./base.txt


pytest 
pytest-cov 
pytest-django 
pytest-dotenv 

 


django-debug-toolbar 